+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [v0.21.1] - 2026-08-21
+
 ### Fixed
 
 - 🐛(backend) fix malware re-analysis of a file already scanned
@@ -502,7 +504,8 @@ and this project adheres to
 - 🌐(front) add english translation for rename modal
 - 🐛(global) fix wrong Content-Type on specific s3 implementations
 
-[unreleased]: https://github.com/suitenumerique/drive/compare/v0.21.0...main
+[unreleased]: https://github.com/suitenumerique/drive/compare/v0.21.1...main
+[v0.21.1]: https://github.com/suitenumerique/drive/releases/v0.21.1
 [v0.21.0]: https://github.com/suitenumerique/drive/releases/v0.21.0
 [v0.20.0]: https://github.com/suitenumerique/drive/releases/v0.20.0
 [v0.19.0]: https://github.com/suitenumerique/drive/releases/v0.19.0